Use non-uniform target widths when refining the grid in gwpcr.mixture.
This accounts for the fact that when mixing poissons, the ones for
a small lambda have a small variance, and hence need a fine grid. 
gwpcr.mixture now takes a grid.with.fun which can choose a maximum
grid with for different values of lambda.

refince_c now takes a vector argument for the width, reusing elements
as necessary.