Replaced clique.size parameter by a more general "loss expression".
This allows more complicated relationships between the raw probability
of an UMI having fewer than T reads (called p0), and the loss, i.e. the
percentage of missed UMIs. This is relevant for designs where the two
strands of a template molecule produce related (reciprocal) UMIs, and
were we thus want to filter based on the existance of the corresponding
UMIs. The loss can then depend in various ways on p0, when we require
both mates to have at least T reads, it e.g. becomes 1 - (1 - p0)^2.

Currently, when evaluating the loss expression, the estimated raw
p0, efficiency and lambda0  can be used.