Work around a limitation in data.table's rbindlist() for factors with many levels.
rbindlist() seems to concatenate all the factor levels together, and
to bail out if the resulting vector is too long (> 2^31 or so). The
previous coding in gwpcrpois.groupest.raw() was prone to triggering
that case, because factor columns retained their levels within each
processed group. rbindlist() thus had to process many rows, each of
which contained only a single actual value, but for factor columns
still all the original levels. The problem is worked around now by
stripping factor columns of the superfluous levels.

Note that should data.table ever gain the ability to detect that all
components indeed refer to the *same* levels vector, the new code
is much less efficient than the olde code. In that case, we should
skip the level pruning if data.table is new enough.